requirements.txt
requests
beautifulsoup4
pandas
yagmail
python-dotenv
